[refactor] extract custom link
